requirements.txt
torch
clip-by-openai
Pillow